Converts a time value given as number of seconds since midnight to the H:M:S format

This function is very similar to strftime with the %X conversion specification. Hour values larger than 24 are permitted. Fractions will be rounded down to the next integer. Non-numeric values are coerced to NA with a warning.

seconds.to.hms(x)

Arguments

x
A (vector of) numbers.

Value

A (vector of) character values of the same length as x.

See also

base::strftime()

Other deprecated functions: coalesce.na-deprecated, df_to_list-deprecated, hms.to.seconds-deprecated, kimisc-deprecated, list_to_df-deprecated, nc-deprecated, nlist-deprecated, ofactor-deprecated, sample.rows-deprecated, thisfile-deprecated, tll-deprecated, vswitch-deprecated

Examples

seconds.to.hms(c(1, 60, 3600.5))

#> `seconds.to.hms()` is deprecated, use `hms::as.character()()` instead.

#> [1] "00:00:01" "00:01:00" "01:00:00"

seconds.to.hms(c(100000, -4000.5))

#> [1] "27:46:40"  "-01:06:41"

seconds.to.hms("invalid")

#> Warning: NAs introduced by coercion

#> [1] NA