Transposes a list of lists

The argument is assumed to be a list of \(n\) (named) lists with length \(m\) each. It is converted to a (named) list of \(m\) elements with length \(n\) each.

Usage

tll(l)

Arguments

l

List of lists, possibly named.

Value

A list of lists corresponding to a transposition of the argument.

See also

Examples

tll(list(list(1, 2), list(3, 4)))
#> `tll()` is deprecated, use `purrr::transpose()` instead.
#> [[1]]
#> [[1]][[1]]
#> [1] 1
#> 
#> [[1]][[2]]
#> [1] 3
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] 2
#> 
#> [[2]][[2]]
#> [1] 4
#> 
#> 
tll(list(list(a = 1, b = 2), list(a = 3, b = 4)))
#> $a
#> $a[[1]]
#> [1] 1
#> 
#> $a[[2]]
#> [1] 3
#> 
#> 
#> $b
#> $b[[1]]
#> [1] 2
#> 
#> $b[[2]]
#> [1] 4
#> 
#> 
tll(list(x = list(a = 1, b = 2), y = list(a = 3, b = 4)))
#> $a
#> $a$x
#> [1] 1
#> 
#> $a$y
#> [1] 3
#> 
#> 
#> $b
#> $b$x
#> [1] 2
#> 
#> $b$y
#> [1] 4
#> 
#>